6/7/22
removing repetition from Voter class by creating functions to run calculations. current result is always 75, but once the second instance of 1 in the random.randint() functions on lines 133 and 135 is removed and replaced with some math, it should place votes that remain in the same corner as the parent while moving around slightly.